TATA TECH data scraping
Using Selenium in VS Code to scrape TATA TECH BSE historical data. The driver initially calls the homepage of finance.yahoo.com and searches "TATE". The script then scrapes the data. There is an unsuccessful attempt to save the data in a csv file. It will further be worked upon in upcoming versions.