Update AQI_Prediction
This python code is used to download the webpages in html format to collect data from the given 'url'. In addition to the download, it also creates directory path if it does not exists. The time duration which is required to download the files is printed at the end of the code which has been done using simple 'time' library.